chore: sort out types
